requirements.txt
tfaip==1.2.6
absl-py<0.11.0
connected-components-3d
datasets
dicom2nifti
editdistance
imgaug
imagecorruptions
imageio
JPype1
libsass
matplotlib!=3.4.3
nibabel
num2words
opencv-python-headless
pandas
pdf_reports
pillow
prettytable
pycocotools
pydicom
python-Levenshtein
PyYAML
sentencepiece
scipy
scikit-learn
seqeval
tensorflow-datasets~=3.0.0
tensorflow_probability
transformers>=4.0.0
treepoem
datasets
